+++ vms/platformvm/metrics/no_op.go
@@ -35,11 +35,11 @@ func (noopMetrics) IncValidatorSetsCached() {}
 
 func (noopMetrics) AddValidatorSetsDuration(time.Duration) {}
 
-func (noopMetrics) AddValidatorSetsHeightDiff(float64) {}
+func (noopMetrics) AddValidatorSetsHeightDiff(uint64) {}
 
-func (noopMetrics) SetLocalStake(float64) {}
+func (noopMetrics) SetLocalStake(uint64) {}
 
-func (noopMetrics) SetTotalStake(float64) {}
+func (noopMetrics) SetTotalStake(uint64) {}
 
 func (noopMetrics) SetSubnetPercentConnected(ids.ID, float64) {}
